==> analyses/tp53_nf1_score/02-qc-rna_expression_score.Rmd <==
---
title: "QC correlation of classifier score with RNA-Seq expression"
author: "K S Gaonkar (D3B)"
output: html_notebook
params:
  base_run:
    label: "1/0 to run with base histology"
    value: 0
    input: integer
  ci_run:
    label: "1/0 to run in CI"
    value: 1
    input: integer
---
    
In this notebook we will find correlation between RNA-Seq and classifier score
We will check if the values are inversely correlated as we would expect?

### Setup
```{r}
library("ggpubr")
library("ggthemes")
library("tidyverse")
library("broom")

# rootdir
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))
data_dir <- file.path(root_dir, "data")
results_dir <- "results"

if ( params$base_run ==0 ){
  clinical<-read.delim(file.path(data_dir,"histologies.tsv"), stringsAsFactors = FALSE)
} else{
  clinical<-read.delim(file.path(data_dir,"histologies-base.tsv"), stringsAsFactors = FALSE)  
}
```

### Classifier scores

 Read in scores from tp53-nf1-classifier 
```{r}

classifier_files <- list.files(path = file.path(results_dir), pattern="*classifier_scores.tsv")

score_df <- data.frame()

if(length(classifier_files)<2)
  stop("there are not two or more files to combine")

for (file in classifier_files) {
 df <- read_tsv(file.path(results_dir, file)) %>%
                dplyr::select(sample_id, tp53_score, tp53_shuffle) %>%
   as.data.frame()
score_df <- bind_rows(score_df, df)
} 

score_df %>%
  write_tsv(file.path(results_dir, "combined_scores.tsv"))
```


### Expression matrices

Read in expression matrices and filter expression matrices to only contain cohort and samples of interest using histology file

```{r}

exp_file <- file.path(data_dir,"gene-expression-rsem-tpm-collapsed.rds")
expressionMatrix<-readRDS(exp_file) 

clinical_filtered <- clinical %>% 
  filter(Kids_First_Biospecimen_ID %in% names(expressionMatrix))

if ( params$ci_run == 1 ){
# Generate RNA library list that can be used to iterate 
rna_library_list <- clinical_filtered %>% pull(RNA_library) %>% unique()

} else{
rna_library_list <- c("stranded", "poly-A")
}

# filter to samples of a particular RNA library type
expression_matrix_list <- lapply(rna_library_list, function(x){
  
  # filter to the samples of interest
  samples_library <- clinical_filtered %>% 
    dplyr::filter(RNA_library == x) %>%
    dplyr::pull(Kids_First_Biospecimen_ID)
  
  # seperate exptression matrix to sample of interest 
  expression_matrix_library <- expressionMatrix %>% dplyr::select(samples_library) %>%
    rownames_to_column()
})
  
```


Format expression matrices to long format

```{r}

expression_mat_to_long <- function(expressionMatrix,score_df,gene){
# use filtered expression matrix for expression based filtering for less than given threshold
# Taking snippet from @jaclyn-taroni idea used in expression based fusion call filtering

expression_long_df <- expressionMatrix %>%
  # rename rowname to GeneSymbol
  dplyr::rename(GeneSymbol=rowname) %>%
  # filter to gene of interest
  dplyr::filter(GeneSymbol == gene) %>%
  # Get the data into long format
  reshape2::melt(variable.name = "Kids_First_Biospecimen_ID",
                 value.name = "expression_value") %>%
  dplyr::mutate(Kids_First_Biospecimen_ID= as.character(Kids_First_Biospecimen_ID)) 

# merge score and expression file
merge_score_exp <- expression_long_df %>%
  left_join(score_df, by=c("Kids_First_Biospecimen_ID"="sample_id"))

return(merge_score_exp)
}

scores_exp_list <- lapply(expression_matrix_list, function(y){
  expression_mat_to_long(y,score_df,"TP53")
})

```


### Correlation between RNA-expression and classifier sccore

```{r}

for (i in 1:length(scores_exp_list)){
  # get each scores file
  scores_file <- scores_exp_list[[i]]
  
  # get the name of the library 
  library_name <- rna_library_list[[i]]
  
  # print out the score
  score_plot <- ggplot(scores_file,aes(expression_value,tp53_score))+
  geom_point() +
  geom_smooth(method='lm') +
  ggtitle(paste0(library_name, " RNA-Seq expression v/s score correlation"))
  
  print(score_plot)
}

  
```


### Corr.test 

```{r}

for (j in 1:length(scores_exp_list)){
  # subset to the scores df
  exp_score <- scores_exp_list[[j]]
  # print out the correlation 
  print(tidy(cor.test(exp_score$expression_value,exp_score$tp53_score))) 
}
  
```


stranded TP53 expression to classifier_score has a 
negative correlation estimate -0.21 at p value 9.816063e-17	

polya TP53 expression to classifier_score has a negligible 
negative correlation estimate -0.095 at p value 0.1330592

polya stranded TP53 expression to classifier_score has a 
negative correlation estimate -0.326 at p value 5.642812e-20	

Overall distribution of expression and classifier score shows 
negative correlation. For stranded and polya stranded, there are statistically
significant negative correlation of -0.21 and -0.326. For these two methods. 
it seemed like expression/classifier score is correlated with functionality/
phenotype of TP53 inactivation in samples. For polyA samples, there is a negative
correlation estimate of -0.095 at p value 0.1330592. And we cannot directly 
use the expression/classifier score to infer functionality/phenotype of TP53 
inactivation in samples

### Sample_id matched polya and stranded

```{r}

sample_id_multi_library <- clinical %>%
  filter(experimental_strategy=="RNA-Seq") %>%
  group_by(sample_id) %>%
  summarise(counts = length(unique(RNA_library)),
            RNA_library = toString(RNA_library)) %>%
  filter(counts > 1)

sample_id_multi_library %>% 
  dplyr::filter(!sample_id %in% c("01", "02", "03", "04", "06", "09", "11", "40")) %>%
  knitr::kable()

```

Note that for sample_id of 01,03,04,09 - a lot of samples exist - indicating issues with indexing of those sample_ids at this point. 
This will likely be fixed later on but at this moment, we will simply remove those. 
```{r}
sample_id_multi_library
```

7316-85 has multiple stranded let's look at the values
```{r}
polya_samples <- clinical_filtered %>%
  dplyr::filter(RNA_library == "poly-A") %>% 
  pull(Kids_First_Biospecimen_ID) %>% 
  unique()

stranded_samples <- clinical_filtered %>%
  dplyr::filter(RNA_library == "stranded") %>% 
  pull(Kids_First_Biospecimen_ID) %>% 
  unique()
  
score_polya_df<- score_df %>% 
  dplyr::filter(sample_id %in% polya_samples)

score_stranded_df<- score_df %>% 
  dplyr::filter(sample_id %in% stranded_samples)

clinical %>%
  left_join(score_stranded_df , by=c("Kids_First_Biospecimen_ID"="sample_id")) %>%
  left_join(score_polya_df,by=c("Kids_First_Biospecimen_ID"="sample_id"),
          suffix = c("_stranded","_polya")) %>%
  filter(sample_id == "7316-85",experimental_strategy=="RNA-Seq") %>%
  select(sample_id,tp53_score_stranded,tp53_score_polya) %>%
  knitr::kable()
```



```{r}

sample_id_multi_library_scores_df <- clinical %>%
  left_join(score_stranded_df , by=c("Kids_First_Biospecimen_ID"="sample_id")) %>%
  left_join(score_polya_df,by=c("Kids_First_Biospecimen_ID"="sample_id"),
            suffix = c("_stranded","_polya")) %>%
  # gather data for sample_ids which have multiple RNA_library
  filter(sample_id %in% sample_id_multi_library$sample_id,
         experimental_strategy=="RNA-Seq") %>%
  group_by(sample_id) %>%
  # because 7316-85 has multiple stranded I'm taking a mean here 
  summarise( tp53_score_stranded = mean(tp53_score_stranded[!is.na(tp53_score_stranded)]),
             tp53_score_polya = mean(tp53_score_polya[!is.na(tp53_score_polya)])) %>%
  reshape2::melt()

  
```


Plotting the scores
```{r}

ggplot(sample_id_multi_library_scores_df,
       aes(y=value,x=as.numeric(factor(sample_id_multi_library_scores_df$variable)))) +
  geom_point()+
  geom_line(aes(color=sample_id)) +
  xlim(c("stranded",
         "polya")) + 
  xlab("RNA_library") + ylab("tp53_score")
```

We could consider utilizing batch correction once completed via [#919](https://github.com/AlexsLemonade/OpenPBTA-analysis/issues/919) to rerun the classifier for this dataset. We can then check whether the scores would be more similar for polyA and stranded. Instead of choosing one over the other without some additional investigation, perhaps we take an average. For the most part, while there is a bias, the scores agree and for 7316-161, although the polyA score is just under 0.5, it is still considered a loss via evidence. Once we investigate all of the evidence, we might consider a re-thresholding of the scores for what determines inactivation/oncogenic TP53 (discussed earlier).
